refactor: post user 컨트롤러를 command query로 분리
