+++ src/main/java/org/sopt/domain/post/presentation/controller/PostQueryController.java
@@ -1,60 +1,31 @@
 package org.sopt.domain.post.presentation.controller;
 
-import java.util.List;
-
-import org.sopt.domain.post.application.service.PostCommandService;
-import org.sopt.domain.post.application.service.PostQueryService;
-import org.sopt.domain.post.application.dto.CreatePostCommand;
 import org.sopt.domain.post.application.dto.PostPageResult;
 import org.sopt.domain.post.application.dto.PostResult;
-import org.sopt.domain.post.application.dto.UpdatePostCommand;
+import org.sopt.domain.post.application.service.PostQueryService;
 import org.sopt.domain.post.domain.model.BoardType;
 import org.sopt.domain.post.presentation.code.PostSuccessCode;
-import org.sopt.domain.post.presentation.dto.request.CreatePostRequest;
 import org.sopt.domain.post.presentation.dto.request.GetPostsRequest;
-import org.sopt.domain.post.presentation.dto.request.UpdatePostRequest;
 import org.sopt.domain.post.presentation.dto.response.PostPageResponse;
 import org.sopt.domain.post.presentation.dto.response.PostResponse;
 import org.sopt.global.response.CommonApiResponse;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PatchMapping;
 import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
+import java.util.List;
+
 @RestController
 @RequestMapping("/posts")
-public class PostController {
-    private final PostCommandService postCommandService;
+public class PostQueryController {
+
     private final PostQueryService postQueryService;
 
-    public PostController(PostCommandService postCommandService, PostQueryService postQueryService) {
-        this.postCommandService = postCommandService;
+    public PostQueryController(PostQueryService postQueryService) {
         this.postQueryService = postQueryService;
-    }
-
-    @PostMapping
-    public ResponseEntity<CommonApiResponse<PostResponse>> createPost(@RequestBody CreatePostRequest request) {
-        request.validate();
-        PostResult result = postCommandService.createPost(new CreatePostCommand(
-                request.boardType(),
-                request.title(),
-                request.content(),
-                request.authorUserId()
-        ));
-        return CommonApiResponse.success(PostSuccessCode.POST_CREATED, new PostResponse(
-                result.id(),
-                result.boardType(),
-                result.title(),
-                result.content(),
-                result.author(),
-                result.createdAt()
-        ));
     }
 
     @GetMapping
@@ -98,18 +69,5 @@ public class PostController {
                 result.author(),
                 result.createdAt()
         ));
-    }
-
-    @PatchMapping("/{postId}")
-    public ResponseEntity<CommonApiResponse<Void>> updatePost(@PathVariable Long postId, @RequestBody UpdatePostRequest request) {
-        request.validate();
-        postCommandService.updatePost(postId, new UpdatePostCommand(request.title(), request.content()));
-        return CommonApiResponse.success(PostSuccessCode.POST_UPDATED, null);
-    }
-
-    @DeleteMapping("/{postId}")
-    public ResponseEntity<CommonApiResponse<Void>> deletePost(@PathVariable Long postId) {
-        postCommandService.deletePost(postId);
-        return CommonApiResponse.success(PostSuccessCode.POST_DELETED, null);
     }
 }
